date place API 개편
1. 데이트 장소 5개로 나누어진 걸 하나의 테이블로  통합
1-1. 쿼리 성능 및 쿼리 단순화를 위해 통합

2. date-place GET 완성
