+++ backend/src/index.ts
@@ -94,6 +94,6 @@ export const API_ROOT = process.env.API_ROOT || `http://localhost:${port}`;
 
 // await job.fetchRestaurant.invoke();
 // await job.fetchTouristSpot.invoke();
-// await job.fetchCulture.invoke();
-await job.fetchSports.invoke();
+await job.fetchCulture.invoke();
+// await job.fetchSports.invoke();
 // await job.fetchShopping.invoke();
