--- backend/src/models/restaurantImage.model.ts
@@ -1,48 +1,0 @@
-import { DataTypes, Model, literal } from "sequelize";
-import { CreationOptional, InferAttributes, InferCreationAttributes } from "sequelize/types/model";
-
-import sequelize, { applyDateHook } from "./index.js";
-import { Restaurant } from "./restaurant.model.js";
-
-export class RestaurantImage extends Model<InferAttributes<RestaurantImage>, InferCreationAttributes<RestaurantImage>> {
-  declare restaurantImageId: CreationOptional<number>;
-  declare contentId: string;
-  declare path: string;
-  declare createdTime: CreationOptional<Date>;
-}
-
-RestaurantImage.init(
-  {
-    restaurantImageId: {
-      field: "restaurant_image_id",
-      type: DataTypes.INTEGER.UNSIGNED,
-      autoIncrement: true,
-      primaryKey: true
-    },
-    contentId: {
-      field: "content_id",
-      type: DataTypes.STRING,
-      allowNull: false,
-      references: {
-        model: Restaurant,
-        key: "contentId"
-      }
-    },
-    path: {
-      type: DataTypes.STRING(200),
-      allowNull: false
-    },
-    createdTime: {
-      field: "created_time",
-      type: "TIMESTAMP",
-      defaultValue: literal("CURRENT_TIMESTAMP")
-    }
-  },
-  {
-    sequelize: sequelize,
-    tableName: "restaurant_image",
-    timestamps: false
-  }
-);
-
-applyDateHook(RestaurantImage);
