+++ backend/src/routes/admin.route.ts
@@ -18,16 +18,8 @@ import { AlbumImage } from "../models/albumImage.model.js";
 import { InquiryImage } from "../models/inquiryImage.model.js";
 import { ContentType } from "../models/contentType.model.js";
 import { RegionCode } from "../models/regionCode.model.js";
-import { Restaurant } from "../models/restaurant.model.js";
-import { TouristSpot } from "../models/touristSpot.model.js";
-import { Culture } from "../models/culture.model.js";
-import { Sports } from "../models/sports.model.js";
-import { Shopping } from "../models/shopping.model.js";
-import { RestaurantImage } from "../models/restaurantImage.model.js";
-import { TouristSpotImage } from "../models/touristSpotImage.model.js";
-import { CultureImage } from "../models/cultureImage.model.js";
-import { SportsImage } from "../models/sportsImage.model.js";
-import { ShoppingImage } from "../models/shoppingImage.model.js";
+import { DatePlace } from "../models/datePlace.model.js";
+import { DatePlaceImage } from "../models/datePlaceImage.model.js";
 
 import { checkPassword } from "../utils/password.util.js";
 import RegionCodeController from "../controllers/regionCode.controller.js";
@@ -194,43 +186,11 @@ const adminOptions: AdminJSOptions = {
       }
     },
     {
-      resource: Restaurant,
+      resource: DatePlace,
       options: actionOptions
     },
     {
-      resource: RestaurantImage,
-      options: actionOptions
-    },
-    {
-      resource: TouristSpot,
-      options: actionOptions
-    },
-    {
-      resource: TouristSpotImage,
-      options: actionOptions
-    },
-    {
-      resource: Culture,
-      options: actionOptions
-    },
-    {
-      resource: CultureImage,
-      options: actionOptions
-    },
-    {
-      resource: Sports,
-      options: actionOptions
-    },
-    {
-      resource: SportsImage,
-      options: actionOptions
-    },
-    {
-      resource: Shopping,
-      options: actionOptions
-    },
-    {
-      resource: ShoppingImage,
+      resource: DatePlaceImage,
       options: actionOptions
     }
   ]
