+++ backend/src/routes/index.ts
@@ -7,6 +7,7 @@ import albumRouter from "./album.route.js";
 import calendarRouter from "./calendar.route.js";
 import inquireRouter from "./inquiry.route.js";
 import noticeRouter from "./notice.route.js";
+import datePlaceRouter from "./datePlace.route.js";
 
 import userAdminRouter from "./user.admin.route.js";
 import coupleAdminRouter from "./couple.admin.route.js";
@@ -16,11 +17,6 @@ import InquireAdminRouter from "./inquire.admin.route.js";
 import InquireImageAdminRouter from "./inquireImage.admin.route.js";
 import SolutionAdminRouter from "./solution.admin.route.js";
 import SolutionImageAdminRouter from "./solutionImage.admin.route.js";
-import restaurantAdminRouter from "./restaurant.admin.js";
-import cultureAdminRouter from "./culture.admin.js";
-import shoppingAdminRouter from "./shopping.admin.js";
-import sportsAdminRouter from "./sports.admin.js";
-import touristSpotAdminRouter from "./touristSpot.admin.js";
 
 import authMiddleware from "../middlewares/auth.middleware.js";
 import checkContentType from "../middlewares/contentType.middleware.js";
@@ -35,6 +31,7 @@ router.use("/album", checkContentType, authMiddleware, albumRouter);
 router.use("/calendar", checkContentType, authMiddleware, calendarRouter);
 router.use("/inquiry", checkContentType, authMiddleware, inquireRouter);
 router.use("/notice", authMiddleware, noticeRouter);
+router.use("/date-place", authMiddleware, datePlaceRouter);
 
 // router.use("/admin/user", authMiddleware, userAdminRouter);
 // router.use("/admin/couple", authMiddleware, coupleAdminRouter);
