+++ backend/src/services/favorite.service.ts
@@ -1,19 +1,14 @@
-import { Transaction, WhereOptions } from "sequelize";
+import { Transaction } from "sequelize";
 
 import { Service } from "./service.js";
-import { ContentType } from "../models/contentType.model.js";
 
-class ContentTypeService extends Service {
+class FavoriteService extends Service {
   getURL(...args: any[]): string {
     throw new Error("Method not implemented.");
   }
 
-  async select(where: WhereOptions<ContentType>): Promise<ContentType | null> {
-    const contentType = await ContentType.findOne({
-      where
-    });
-
-    return contentType;
+  select(...args: any[]): Promise<any> {
+    throw new Error("Method not implemented.");
   }
 
   create(transaction: Transaction | null, ...args: any[]): Promise<any> {
@@ -29,4 +24,4 @@ class ContentTypeService extends Service {
   }
 }
 
-export default ContentTypeService;
+export default FavoriteService;
